Fixing Windows path handling for synthetic datasets
If a dataset is generated with Unity running on a Windows OS, the
resultant paths in the Captures.json records will have Windows path
separators (i.e. '\\') whereas the synthetic dataset handler expects
these paths to be posix-style ('/').

Added a _fix_filenames step to replace any Windows-style directory
separators in the captures' filenames before the clean-up step checks
for the path's existence.

Tested manually in a kubeflow pipeline with Windows-generated data